Force use of python 2 for SCons build and execution
SCons does not work on Python 3 at the moment. In order
for the stack to be built and run using python 3 we
need to ensure that SCons itself does not use the
default python that is on the stack. For this reason
we hardwire python2.7 into the shebang line and into
the eupspkg.cfg.sh file.

This requires that packages that introspect python in
SConstruct files must now call out to the python on the
PATH rather than assuming that SCons python is the
relevant python.